debug: free meals form navigation
